Rename compiler/SourceBuilder.ts to compiler/source-builder.ts
