+++ src/compiler/source-builder.ts
